✨ Feat(floating_terminal): add callback function
Implements a new callback function that allows interaction with the
result of commands run in the floating terminal window

BREAKING CHANGE: The ispreupdate option is now deprecated and will be
removed in a future release.

The callback function option provides a better way to trigger subsequent
functions making ispreupdate obsolete.