✨ Feat(update-workflow): improve user experience when update fails
When an update fails, the user will be prompted to remove the build dir
and try again.

This addresses the main cause of failure for subsequent updates in an
existing source directory.

Better use of the last_update.count value: Failed updates will set the
count to "?" as will removing the source dir. Count is not set to 0
until after the update completes successfully.